backend/scripts/repo_A/requirements.txt
chardet
mpu
numpy
openpyxl
pandas
PyYAML
scikit-learn
scipy
